requirements.txt
Django==1.5.4
csvImporter==0.1.3.2
django-adaptors==0.2.5
lxml==3.3.5
oauthlib==0.6.3
requests==2.1.0
requests-oauthlib==0.4.0
twython==3.1.2
twython-django==1.5.2
